+++ packages/at_client/lib/src/rpc/at_rpc.dart
@@ -28,6 +28,7 @@ abstract class AtRpcCallbacks {
 /// - Responder:
 /// ```
 /// ```
+@experimental
 class AtRpc {
   static final AtSignLogger logger = AtSignLogger('AtRpc');
 
@@ -145,6 +146,7 @@ class AtRpc {
             NotificationParams.forUpdate(requestRecordID, value: requestJson),
             checkForFinalDeliveryStatus: false,
             waitForFinalDeliveryStatus: false);
+        sent = true;
         logger.info('Notification ${requestRecordID.toString()} sent');
       } catch (e) {
         if (attemptNumber < maxSendAttempts) {