+++ packages/at_client_mobile/example/lib/src/screens/login.dart
@@ -3,6 +3,7 @@ import 'package:at_client_mobile_example/src/services/sdk.service.dart';
 import 'package:flutter/material.dart';
 import 'package:at_onboarding_flutter/at_onboarding_flutter.dart';
 import 'package:at_app_flutter/at_app_flutter.dart' show AtEnv;
+import 'package:at_client_mobile/at_client_mobile.dart';
 
 class LoginScreen extends StatefulWidget {
   /// Login screen id.