+++ packages/at_client_mobile/lib/src/keychain_manager.dart
@@ -27,6 +27,10 @@ class KeyChainManager {
   KeyChainManager._internal();
 
   factory KeyChainManager.getInstance() {
+    if (Platform.isWindows) {
+      // ignore: undefined_method
+      Win32BiometricStoragePlugin.registerWith();
+    }
     return _singleton;
   }